Skeleton code for payload extraction application
