+++ pyproject.toml
@@ -58,6 +58,7 @@ include-package-data = true
 eve_json = "suricatalog.scripts.eve_json:main"
 eve_log = "suricatalog.scripts.eve_log:main"
 eve_server = "suricatalog.scripts.eve_server:main"
+eve_payload = "suricatalog.scripts.eve_payload:main"
