+++ test/test_payload.py
@@ -1,0 +1,13 @@
+"""
+Unit test code
+"""
+import unittest
+
+
+class PayloadTestCase(unittest.TestCase):
+    def test_something(self):
+        self.assertEqual(True, False)  # add assertion here
+
+
+if __name__ == '__main__':
+    unittest.main()
